==> scgpt/02_seurat2scanpy.ipynb <==
import scanpy as sc
import numpy as np

import anndata2ri

# Activate the anndata2ri conversion between SingleCellExperiment and AnnData
anndata2ri.activate()

#Loading the rpy2 extension enables cell magic to be used
#This runs R code in jupyter notebook cells
%load_ext rpy2.ipython

sc.settings.verbosity = 3
sc.logging.print_versions()

%%R
suppressPackageStartupMessages(library(Seurat))

sc_seurat_1 <- readRDS("../human_data/meclo_ht2-ht6_seurat_table_20240309_labelled-mimeticsonly_preprocessed.rds")

sc_seurat_1

%%R
BiocManager::install("SingleCellExperiment")

%%R -o sc_ann_1
#convert the Seurat object to a SingleCellExperiment object
sc_ann_1 <- as.SingleCellExperiment(sc_seurat_1)

sc_ann_1

sc_ann_1.obs

sc_ann_1.write('../human_data/meclo_ht2-ht6_seurat_table_20240309_labelled-mimeticsonly_preprocessed.h5ad')

## Mouse and Zebrafish data (new)

import scanpy as sc
import numpy as np

import anndata2ri

# Activate the anndata2ri conversion between SingleCellExperiment and AnnData
anndata2ri.activate()

#Loading the rpy2 extension enables cell magic to be used
#This runs R code in jupyter notebook cells
%load_ext rpy2.ipython

sc.settings.verbosity = 3
sc.logging.print_versions()

%%R
suppressPackageStartupMessages(library(Seurat))

sc_seurat <- readRDS("../mouse/meclo_seurat_table_mouse-humanlabels_mimeticsonly-20240306_cleaned.rds")

sc_seurat

%%R -o sc_ann
#convert the Seurat object to a SingleCellExperiment object
sc_ann <- as.SingleCellExperiment(sc_seurat)

sc_ann

sc_ann

sc_ann.write('../mouse/meclo_seurat_table_mouse-humanlabels_mimeticsonly-20240306_cleaned.h5ad')

%%R
suppressPackageStartupMessages(library(Seurat))

sc_seurat <- readRDS("../zebrafish/zebrafish_mec_aftercut5-humanlabels_mimeticsonly-20240306_cleaned.rds")

sc_seurat

%%R -o sc_ann
#convert the Seurat object to a SingleCellExperiment object
sc_ann <- as.SingleCellExperiment(sc_seurat)

sc_ann

sc_ann

sc_ann.write('../zebrafish/zebrafish_mec_aftercut5-humanlabels_mimeticsonly-20240306_cleaned.h5ad')
